Move error classifier to internal/mirror
- Classifier contains application-level advice (CLI flags, docs links) specific to mirror commands
- Document classifier placement convention in README and doc.go
