+++ internal/mirror/errdiag/classify_network_test.go
